ci: 🔧 update trigger for release action (#35)
* ci: 🔧 update trigger for release action

Commented out the push trigger for the master branch and improved comments in the semantic-release step for clarity. The workflow can now be triggered manually via workflow_dispatch.

* Set author info for semantic release commits

Adds environment variables to configure the author and committer name/email for semantic release commits, ensuring they are attributed to the GitHub Actions bot.

---------

Co-authored-by: Konda Sandeep <139792713+SandeepOCT@users.noreply.github.com>